Fixing clippy.

+++ crates/db_schema/src/impls/community.rs
@@ -282,11 +282,11 @@ impl Community {
   pub async fn update_federated_followers(
     pool: &mut DbPool<'_>,
     for_community_id: CommunityId,
-    new_subscribers: i64,
+    new_subscribers: i32,
   ) -> LemmyResult<Self> {
     let conn = &mut get_conn(pool).await?;
     diesel::update(community::table.find(for_community_id))
-      .set(community::dsl::subscribers.eq(new_subscribers as i32))
+      .set(community::dsl::subscribers.eq(new_subscribers))
       .get_result(conn)
       .await
       .with_lemmy_type(LemmyErrorType::CouldntUpdateCommunity)
